fitSpliced {OpVaR}	R Documentation

Estimation of the threshold, the body and the tail parameters for a spliced distribution

Description

Given a dataset with a chosen distribution for the data in the body and another distribution in the tail, the threshold, the parameters of the body and the tail distributions and the weights are estimated.

Usage

  fitSpliced(cell, body, tail, method, thresh = NULL)

  fitSplicedPar(cell, thresh, body, tail)

Arguments

cell
List containing the data in the component cell$Loss

body
Distribution in the body. Can be chosen between "gamma", "lnorm", "weibull" or "erlang"

tail
Distribution in the tail. Can be chosen between "gpd", "gamma", "lnorm", "weibull" or "gh"

method
Method for the threshold estimation. In case of a GPD tail, it can be chosen between "BestFit", "Fixed", "dAMSE", "danielsson", "DK", "hall", "Himp", "HW", "mindist" or "RT"

thresh
Predetermined threshold quantile (if method="Fixed")

Details

In the spliced model, the distribution of the data is spliced into two distributions, one in the body and another in the tail. The density f(x) for a spliced distribution with threshold τ, weights w, body distribution with density g and cumulative distribution function G and tail distribution with density h and cumulative distribution function H is given as below:

If x≤ τ: f(x) = w* g(x)/G(τ)

If x>τ: f(x) = (1-w)* h(x)/(1-H(τ))

The weight w is required to normalise the density function. The estimation of the spliced distribution consists of three steps. In the first step, the threshold for the selected body and tail distribution is estimated. For further details on the methods for the threshold estimation, see fitThreshold. In the second step, the parameters of the body and the tail distribution are obtained by maximum likelihood estimation. For spliced distributions, truncated body distributions are fitted to truncated data. For the estimation of the parameters of the tail distribution, only the data points above the threshold are used. In the last step, the weight w is obtained.

Value

Returns a sevdist object of type 'spliced' with the given body and tail distributions fitted to the loss data.

Author(s)

Christina Zou, Leonie Wicht

See Also

fitSplicedBestFit, fitThreshold

Examples

 data(lossdat)
 fitSpliced(lossdat[[3]],"gamma","gpd",method="Fixed",thresh=2000)

[Package OpVaR version 1.1.1 Index]