Mixing {OpVaR}	R Documentation

Extended Dynamic Weighted Mixture Model

Description

Implementation of the dynamic weighted mixture model of Frigessi et al., 2001

Usage

  dmixing(x,sevdist,log=FALSE)
  pmixing(q,sevdist,lower.tail=TRUE)
  qmixing(p,sevdist,lower.tail=TRUE)
  rmixing(n,sevdist)

Arguments

x,q
vector of quantiles

p
vector of probabilities.

n
positive integer, number of observations

sevdist
sevdist object

log
logical; if TRUE, probabilities p are given as log(p).

lower.tail
logical; if TRUE (default), probabilities are P[X <= x] otherwise, P[X > x].

Details

The dynamic mixture pdf is given by f(x)=1/r*((1-W(x))*g(x)+W(x)*h(x)), where W(x) denotes a Cauchy cdf, g and h body and tail pdfs, r is a normalizing constant.

Author(s)

Marius Pfeuffer

References

A. Frigessi et al.: A Dynamic Mixture Model for Unsupervised Tail Estimation without Threshold Selection, Extremes 5(3):219-235, 2002

Y. Hu: User's Guide for evmix Package in R., Working Paper, 2013

Examples

  ### Create mixing sevdist object
  sevdist=buildMixingSevdist("weibull", c(2,6), "gpd", c(0,4,-0.5), c(8,.8))
  ### Evaluate pdf
  dmixing(1,sevdist)

[Package OpVaR version 1.1.1 Index]