OpVaR-package {OpVaR}	R Documentation

Statistical Methods for Modelling Operational Risk

Description

Functions for computing the value-at-risk in compound Poisson models. The implementation comprises functions for modeling loss frequencies and loss severities with plain, mixed (Frigessi et al. (2012) <doi:10.1023/A:1024072610684>) or spliced distributions using Maximum Likelihood estimation and Bayesian approaches (Ergashev et al. (2013) <doi:10.21314/JOP.2013.131>). In particular, the parametrization of tail distributions includes the fitting of Tukey-type distributions (Kuo and Headrick (2014) <doi:10.1155/2014/645823>). Furthermore, the package contains the modeling of bivariate dependencies between loss severities and frequencies, Monte Carlo simulation for total loss estimation as well as a closed-form approximation based on Degen (2010) <doi:10.21314/JOP.2010.084> to determine the value-at-risk.

Author(s)

Christina Zou [aut,cre], Marius Pfeuffer [aut], Matthias Fischer [aut], Kristina Dehler [ctb], Nicole Derfuss [ctb], Benedikt Graswald [ctb], Linda Moestel [ctb], Jixuan Wang [ctb], Leonie Wicht [ctb]

Maintainer: Christina Zou <christina.zou@maths.ox.ac.uk>

References

Degen, M. (2010): The calculation of minimum regulatory capital using single-loss approximations. The Journal of Operational Risk, 5(4), 3.

Dehler, K. (2017): Bayesianische Methoden im operationellen Risiko. Master's Thesis, Friedrich-Alexander-University Erlangen-Nuremberg.

Ergashev, B. et al. (2013): A Bayesian Approach to Extreme Value Estimation in Operational Risk Modeling. Journal of Operational Risk 8(4):55-81

Frigessi, A. et al. (2002): A Dynamic Mixture Model for Unsupervised Tail Estimation Without Threshold Selection. Extremes 5(3):219-235

Kuo, T. C. and Headrick, T. C. (2014): Simulating Univariate and Multivariate Tukey g-and-h Distributions Based on the Method of Percentiles. ISRN Probability and Statistics.

Pfaelzner, F. (2017): Einsatz von Tukey-type Verteilungen bei der Quantifizierung von operationellen Risiken. Master's Thesis, Friedrich-Alexander-University Erlangen-Nuremberg.

Reynkens, T. et al. (2017): Modelling Censored Losses Using Splicing: a global fit strategy with mixed Erlang and Extreme Value Distributions. Insurance: Mathematics and Economics 77:67-77

Tukey, J. W. (1960): The Practical Relationship between the Common Transformations of Counts of Amounts. Technical Report 36, Princeton University Statistical Techniques Research Group, Princeton.

[Package OpVaR version 1.1.1 Index]