buildFreqdist {OpVaR}	R Documentation

Building a freqdist object

Description

Building a freqdist object

Usage

  buildFreqdist(freq.distr, freq.param)

Arguments

freq.distr
A string giving the name of the distribution.

Distributions "pois" and "nbinom" are recognised.

freq.param
List of numerical values giving the parameters for the given distribution.

Details

A freqdist object with type 'plain' is generated, i.e. for a given cell the losses occur with frequency given by distribution 'distr' with parameters given by 'param'.

Author(s)

Christina Zou

See Also

The sevdist objects according to different types are built via buildPlainSevdist, buildSplicedSevdist and buildMixingSevdist. For given data, a freqdist object can be fitted using fitFreqdist.

Examples

  # Poisson distributed frequency with lambda = 50
  
  freqdist1 = buildFreqdist("pois", 50)

[Package OpVaR version 1.1.1 Index]