buildMixingSevdist {OpVaR}	R Documentation

Building a dynamic mixture model as a sevdist object

Description

Building a dynamic mixture model as a sevdist object

Usage

  buildMixingSevdist(body.distr, body.param, tail.distr, tail.param, mixing.param)

Arguments

body.distr
A string giving the name of the body distribution.

body.param
Vector of the parameters for the given body distribution.

tail.distr
A string giving the name of the tail distribution.

tail.param
Vector of the parameters for the given tail distribution.

Distributions "lgamma", "weibull", "gpd", "gh" are recognised.

mixing.param
Vector of the parameters for the given mixing distribution.

Details

A sevdist object with type 'mixing' is generated, i.e. the distribution of the severity is a dynamic mixture of the distributions 'body.distr' with parameters given by 'body.param' and 'tail.distr' with parameters given by 'tail.param' using the cdf of a mixing distribution 'mixing.distr' with parameters 'mixing.param' according to Frigessi et al. (2002).

The density f(x) for a dynamic mixture model with body distribution having density g and tail distribution having density h and cumulative distribution function p of the mixing distribution is given below:

f(x) = C*[p(x)*g(x) +(1-p(x))*h(x)]

Author(s)

Christina Zou

References

Frigessi, A., Haug, O., & Rue, H. (2002). A dynamic mixture model for unsupervised tail estimation without threshold selection. Extremes, 5(3), 219-235.

Examples

  # Create mixing sevdist object
  sevdist = buildMixingSevdist("weibull", c(2,6), "gpd", c(0,4,-0.5), c(8,.8))
  # Plot pdf
  plot(sevdist)

[Package OpVaR version 1.1.1 Index]