gpd {OpVaR}	R Documentation

Generalized Pareto Distribution

Description

Density, distribution function, quantile function, and random generation for generalized Pareto distribution.

Usage

dgpd(x, loc, scale, shape, log=FALSE)
pgpd(q, loc, scale, shape) 
qgpd(p, loc, scale, shape) 
rgpd(n, loc, scale, shape) 

Arguments

x, q
vector of quantiles.

p
vector of probabilities.

n
number of observations.

loc
location parameter.

scale
scale parameter.

shape
shape parameter.

log
logical; logarithm of output

Details

Generalized Pareto Distribution with density

f(x)=1/scale * (1 + shape * (x - loc)/scale)^(-1/shape - 1)

Value

dgpd gives the density, pgpd gives the distribution function, qgpd gives the quantile function, and rgpd generates random deviates.

Examples

dgpd(110,100,2,4) # 0.01112233

[Package OpVaR version 1.1.1 Index]