dsevdist {OpVaR}	R Documentation

Evaluating Plain, Spliced or Mixing Severity Distributions

Description

Functions for evaluating pdf, cdf, quantile function and random number generation given a loss severity model (sevdist object)

Usage

dsevdist(x, sevdist)

Arguments

x

sevdist

Author(s)

Marius Pfeuffer

Examples

##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x, sevdist) 
{
    if (sevdist$type == "plain") 
        return(sevdist$par[[1]](x))
    if (sevdist$type == "spliced") 
        return(dspliced(x, sevdist))
    if (sevdist$type == "mixing") 
        return(dmixing(x, sevdist))
  }

[Package OpVaR version 1.1.1 Index]