fitDependency {OpVaR}	R Documentation

Function for fitting bivariate Copulas

Description

Function for fitting bivariate dependency between loss frequencies and severities.

Usage

fitDependency(cell, copula_family, method = "mle")

Arguments

cell
a lossdat cell

copula_family
0: independence, 1: Gaussian, 2: t, 3: Clayton, 4: Gumbel, 5: Frank, 6: Joe, see Vine Copula Package ?BiCop encoding

method
as in fitCopula of the copula package. The method "itau" should not be used as dependencies between a continuous and a discrete random variable are modeled.

Value

A BiCop object

Author(s)

Marius Pfeuffer

References

Schepsmeier et al.: VineCopula package, CRAN

Examples

### Fit Joe Copula
data(lossdat)
fitDependency(lossdat[[1]],6)

[Package OpVaR version 1.1.1 Index]