fitFreqdist {OpVaR}	R Documentation

Fitting the frequency distribution

Description

Fitting the frequency distribution of operational losses

Usage

  fitFreqdist(cell, distr)

Arguments

cell
A data frame giving the losses (cell$Loss) and the user-defined period (cell$Period) in which the loss occured.

distr
The frequency distribution: either "pois" or "nbinom"

Details

First, the number of losses per period is counted. With the results a Poisson or Negative Binomial distribution is fitted for the number of losses per period using a Maximum Likelihood estimator.

Value

Returns a freqdist object with the given distribution fitted to the loss frequency.

Author(s)

Christina Zou

See Also

To build a freqdist object independently, use buildFreqdist.

Examples

  data(lossdat)
  fitFreqdist(lossdat[[1]],"pois")
  fitFreqdist(lossdat[[2]],"nbinom")

[Package OpVaR version 1.1.1 Index]