fitMixing {OpVaR}	R Documentation

Maximum Likelihood Estimation

Description

Maximum Likelihood Estimation for the dynamic weighted mixture model of Frigessi et al., 2002

Usage

fitMixing(cell, body, tail, method="L-BFGS-B", c_location0=0.75, c_scale0=2)

Arguments

cell
lossdat cell

body
body distribution, either gamma, lnorm or weibull

tail
tail distribution, either gamma, lnorm, weibull or gpd

method
optimization method, default is "L-BFGS-B"

c_location0
empirical quantile of loss severity data used for initialization of Cauchy location parameter in optimization: quantile(cell$Loss,c_location0)

c_scale0
scaling factor for empirical standard deviation used for initialization of Cauchy scale parameter in optimization: sd(cell$Loss)/c_scale0

Details

Body and tail parameters are initialized by method of moments estimators. Cauchy location is initialized by empirical 70

Value

Returns a sevdist object of type 'mixing' with the given body and tail distributions fitted to the loss data.

Author(s)

Marius Pfeuffer

References

Frigessi et al. A Dynamic Mixture Model for Unsupervised Tail Estimation without Threshold Selection, Extremes 5(3):219-235, 2003

Examples

data(lossdat)
sev=fitMixing(lossdat[[1]],"weibull","gpd")
sev
plot(sev,5000)

[Package OpVaR version 1.1.1 Index]