fitWeights {OpVaR}	R Documentation

Fitting the weights of the body and the tail for a spliced distribution

Description

Fitting the weights of the spliced distribution.

Usage

  fitWeights(cell, thresh)

Arguments

cell
List containing the data in the component Loss

thresh
Threshold between the body and the tail.

Details

Given the threshold t, this function fits the weight on the body distribution as part of fitting the parameters of the spliced distribution. The weights depends only on the number of data points below and above the threshold, not on the distribution functions in the body and the tail.

Author(s)

Christina Zou, Leonie Wicht

See Also

fitThreshold, fitSpliced

Examples

    data(lossdat)
    thresh<-quantile(lossdat[[1]]$Loss, 0.9)
    w<-fitWeights(lossdat, thresh)
    
  

[Package OpVaR version 1.1.1 Index]